.history/backend/app/services/audio_service_20250927173324.py
# backend/app/services/audio_service.py
import asyncio
import aiohttp
import aiofiles
import ffmpeg
import subprocess
import os
from typing import Dict, List, Optional
from datetime import datetime
import json
import logging

logger = logging.getLogger(__name__)

class AudioStreamingService:
    def __init__(self):
        self.active_streams: Dict[str, Dict] = {}
        self.current_tracks: Dict[str, Dict] = {}
        self.ad_rotation: Dict[str, List] = {}
        
    async def start_channel_stream(self, channel_id: str, source_url: str):
        """Start HLS stream for a channel"""
        try:
            # Create output directory
            output_dir = f"streams/{channel_id}"
            os.makedirs(output_dir, exist_ok=True)
            
            # HLS stream configuration
            hls_path = f"{output_dir}/stream.m3u8"
            
            # FFmpeg command for HLS streaming
            ffmpeg_cmd = [
                'ffmpeg',
                '-i', source_url,
                '-c:a', 'aac',
                '-b:a', '128k',
                '-f', 'hls',
                '-hls_time', '10',
                '-hls_list_size', '6',
                '-hls_flags', 'delete_segments',
                '-hls_segment_filename', f'{output_dir}/segment_%03d.ts',
                hls_path
            ]
            
            # Start FFmpeg process
            process = await asyncio.create_subprocess_exec(
                *ffmpeg_cmd,
                stdout=asyncio.subprocess.PIPE,
                stderr=asyncio.subprocess.PIPE
            )
            
            self.active_streams[channel_id] = {
                'process': process,
                'hls_url': f"/streams/{channel_id}/stream.m3u8",
                'started_at': datetime.now(),
                'listeners': 0
            }
            
            logger.info(f"Started HLS stream for channel {channel_id}")
            return f"/streams/{channel_id}/stream.m3u8"
            
        except Exception as e:
            logger.error(f"Error starting stream for {channel_id}: {e}")
            raise

    async def play_track(self, channel_id: str, track_data: Dict):
        """Play a specific track on a channel"""
        self.current_tracks[channel_id] = {
            **track_data,
            'started_at': datetime.now(),
            'progress': 0
        }
        
        # Update WebSocket clients
        await self.broadcast_now_playing(channel_id, track_data)

    async def insert_advertisement(self, channel_id: str, ad_data: Dict):
        """Insert advertisement into stream"""
        # Implementation for ad insertion
        pass

    async def broadcast_now_playing(self, channel_id: str, track_data: Dict):
        """Broadcast now playing info to WebSocket clients"""
        from .websocket_manager import websocket_manager
        
        message = {
            'type': 'now_playing',
            'channel_id': channel_id,
            'track': track_data,
            'timestamp': datetime.now().isoformat()
        }
        
        await websocket_manager.broadcast_to_channel(channel_id, message)

class PlaylistScheduler:
    def __init__(self):
        self.schedules: Dict[str, Dict] = {}
        
    async def generate_playlist(self, channel_id: str, tracks: List[Dict]):
        """Generate automated playlist with intelligent scheduling"""
        # Implement smart playlist generation
        # Consider: time of day, listener preferences, artist variety, etc.
        pass

# Global instance
audio_service = AudioStreamingService()