.history/backend/requirements_20250927180218.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aioredis==2.0.1
websockets==12.0
ffmpeg-python==0.2.0
pydub==0.25.1
celery==5.3.4
python-dotenv==1.0.0
aiofiles==23.2.1
